+++ src/modern-async.mjs
@@ -25,8 +25,6 @@ export { default as sleepCancellable } from './sleepCancellable.mjs'
 export { default as sleepPrecise } from './sleepPrecise.mjs'
 export { default as sleepPreciseCancellable } from './sleepPreciseCancellable.mjs'
 export { default as some } from './some.mjs'
-export { default as someLimit } from './someLimit.mjs'
-export { default as someSeries } from './someSeries.mjs'
 export { default as timeout } from './timeout.mjs'
 export { default as TimeoutError } from './TimeoutError.mjs'
 export { default as timeoutPrecise } from './timeoutPrecise.mjs'
