--- src/someSeries.test.mjs
@@ -1,55 +1,0 @@
-
-import { expect, test } from '@jest/globals'
-import someSeries from './someSeries.mjs'
-import Deferred from './Deferred.mjs'
-import { range } from 'itertools'
-
-test('someSeries all no pass', async () => {
-  const callCount = {}
-  ;[...range(3)].forEach((i) => { callCount[i] = 0 })
-  const d = new Deferred()
-  const ds = [...range(3)].map(() => new Deferred())
-  const p = someSeries([...range(3)], async (v, i) => {
-    callCount[i] += 1
-    ds[i].resolve()
-    await d.promise
-    return false
-  })
-  await ds[0].promise
-  expect(callCount[0]).toBe(1)
-  expect(callCount[1]).toBe(0)
-  expect(callCount[2]).toBe(0)
-  d.resolve()
-  const res = await p
-  expect(res).toBe(false)
-  expect(callCount[0]).toBe(1)
-  expect(callCount[1]).toBe(1)
-  expect(callCount[2]).toBe(1)
-})
-
-test('someSeries some pass', async () => {
-  const callCount = {}
-  ;[...range(3)].forEach((i) => { callCount[i] = 0 })
-  const d = new Deferred()
-  const ds = [...range(3)].map(() => new Deferred())
-  const p = someSeries([...range(3)], async (v, i) => {
-    callCount[i] += 1
-    ds[i].resolve()
-    await d.promise
-    if (i === 1) {
-      return true
-    } else {
-      return false
-    }
-  })
-  await ds[0].promise
-  expect(callCount[0]).toBe(1)
-  expect(callCount[1]).toBe(0)
-  expect(callCount[2]).toBe(0)
-  d.resolve()
-  const res = await p
-  expect(res).toBe(true)
-  expect(callCount[0]).toBe(1)
-  expect(callCount[1]).toBe(1)
-  expect(callCount[2]).toBe(0)
-})
